requirements.txt
pyTelegramBotAPI
deep-translator
huggingface_hub
gradio-client
dublib==0.13.0
Pillow
g4f